test string logic finished
Added logic to the test string function and updated the testWindow class so the it will display whether or not the string was accepted or rejected... Hard coded instances of the DFA, states, and transitions for testing purposes. Will need to integrate that into the GUI at a later date.