Working string testing functionality
The test string on automata function now correctly constructs the DFA with the automaton.ts library using the JSON data of the Canvas DFA. With this, input string can now be tested on the constructed DFA. The function correctly identifies if a test string is accepted, rejected, or contains invalid tokens. There are still some inconsistencies with recognizing invalid DFAs.